requirements.txt
plotly=5.16.1
python-dotenv=1.0.0
pandas=2.1.0
scikit-learn=1.3.0
streamlit=1.2.6